move auth helper functions to a separate utils file
