+++ src/app/Scenes/Onboarding/OnboardingCreateAccount/OnboardingCreateAccount.tsx
@@ -4,10 +4,10 @@ import { createStackNavigator, StackScreenProps, TransitionPresets } from "@reac
 import { OnboardingNavigationStack } from "app/Scenes/Onboarding/Onboarding"
 import { OnboardingSocialPick } from "app/Scenes/Onboarding/OnboardingSocialPick"
 import { OnboardingWebView, OnboardingWebViewRoute } from "app/Scenes/Onboarding/OnboardingWebView"
-import { showBlockedAuthError } from "app/store/AuthModel"
 import { GlobalStore } from "app/store/GlobalStore"
 import { BackButton } from "app/system/navigation/BackButton"
 import { ArtsyKeyboardAvoidingView } from "app/utils/ArtsyKeyboardAvoidingView"
+import { showBlockedAuthError } from "app/utils/auth/authHelpers"
 import { useScreenDimensions } from "app/utils/hooks"
 import { FormikProvider, useFormik, useFormikContext } from "formik"
 import React, { useEffect, useRef, useState } from "react"
