+++ src/app/Scenes/Onboarding/OnboardingLogin.tsx
@@ -1,10 +1,10 @@
 import { Spacer, Flex, Box, useColor, Text, Touchable, Button } from "@artsy/palette-mobile"
 import { StackScreenProps } from "@react-navigation/stack"
 import { Input } from "app/Components/Input"
-import { showBlockedAuthError } from "app/store/AuthModel"
 import { GlobalStore } from "app/store/GlobalStore"
 import { BackButton } from "app/system/navigation/BackButton"
 import { ArtsyKeyboardAvoidingView } from "app/utils/ArtsyKeyboardAvoidingView"
+import { showBlockedAuthError } from "app/utils/auth/authHelpers"
 import { useScreenDimensions } from "app/utils/hooks"
 import { FormikProvider, useFormik, useFormikContext } from "formik"
 import React, { useEffect, useRef } from "react"
