+++ src/app/Scenes/Onboarding/OnboardingSocialPick.tsx
@@ -2,12 +2,9 @@ import { EnvelopeIcon, Spacer, Flex, Text, Join, Button, LegacyScreen } from "@a
 import { NavigationProp, useNavigation } from "@react-navigation/native"
 import { captureMessage } from "@sentry/react-native"
 import LoadingModal from "app/Components/Modals/LoadingModal"
-import {
-  AuthPromiseRejectType,
-  AuthPromiseResolveType,
-  showBlockedAuthError,
-} from "app/store/AuthModel"
+import { AuthPromiseRejectType, AuthPromiseResolveType } from "app/store/AuthModel"
 import { GlobalStore } from "app/store/GlobalStore"
+import { showBlockedAuthError } from "app/utils/auth/authHelpers"
 import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
 import { osMajorVersion } from "app/utils/platformUtil"
 import { capitalize } from "lodash"
@@ -43,6 +40,11 @@ export const OnboardingSocialPick: React.FC<OnboardingSocialPickProps> = ({ mode
 
   const handleErrorWithAlternativeProviders = (meta: AuthPromiseRejectType["meta"]) => {
     const titleizedProvider = capitalize(meta?.provider ?? "")
+
+    if (!meta) {
+      return
+    }
+
     const {
       email,
       name,
@@ -51,7 +53,7 @@ export const OnboardingSocialPick: React.FC<OnboardingSocialPickProps> = ({ mode
       oauthToken,
       idToken,
       appleUid,
-    } = meta!
+    } = meta
     const navParams: Omit<
       OnboardingNavigationStack["OnboardingSocialLink"],
       "tokenForProviderToBeLinked"
